+++ src/main/kotlin/com/cycode/plugin/cli/models/CliError.kt
@@ -1,7 +1,9 @@
 package com.cycode.plugin.cli.models
 
 data class CliError(
-    val code: Int,
+    // FIXME(MarshalX): sometimes CLI uses `code` and sometimes `error` for the same thing
+    val code: String? = "Unknown",
+    val error: String? = "Unknown",
     val message: String,
     val softFail: Boolean? = false,
 )